% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_and_push.R
\name{commit_and_push}
\alias{commit_and_push}
\title{Title}
\usage{
commit_and_push()
}
\value{
All changes in projects are staged and pushed to Git
}
\description{
Title
}
\examples{
commit_and_push()
}
