% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folder_structure.R
\name{create_folder_structure}
\alias{create_folder_structure}
\title{Title}
\usage{
create_folder_structure(directory)
}
\arguments{
\item{directory}{}
}
\value{
standard folder structure
}
\description{
Title
}
\examples{
create_folder_structure(getwd())
}
