% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_names.R
\name{s_names}
\alias{s_names}
\title{Simple names}
\usage{
s_names(names)
}
\arguments{
\item{names}{}
}
\value{
A character vector
}
\description{
s_names() removes symbols such as parentheses and colons and
substitute symbols (|, :, and +) with underscores. No letters are capitalized.
}
\examples{
x <- c("x (1)", "X 2", "X - 3", "X:4")
s_names(x)
}
