% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_table.R
\name{s_table}
\alias{s_table}
\title{Simple tables}
\usage{
s_table(df)
}
\arguments{
\item{df}{}
}
\value{
A table in kable format
}
\description{
s_table() converts a input table to a table for visualization
in rmarkdown format.
}
\examples{
x <- data.frame(x = c(1, 2), y = c(4, 2))
s_table(x)
}
