% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scroll_table.R
\name{scroll_table}
\alias{scroll_table}
\title{Scroll tables}
\usage{
scroll_table(dataframe)
}
\arguments{
\item{dataframe}{}
}
\value{
A scroll table
}
\description{
scroll_table() converts a input table to a scroll table.
}
\examples{
x <- data.frame(x = c(1, 2), y = c(4, 2))
scroll_table(x)
}
