% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stat.R
\name{sum_stat}
\alias{sum_stat}
\title{Summary statistics}
\usage{
sum_stat(df, variables)
}
\arguments{
\item{df}{A data.frame with values that needs to be summarised}

\item{variables}{Selected columns from the data.frame}
}
\value{
Minimum, median and maximum values of selected variables
}
\description{
sum_stat returns summary statistics of selected variables
from a data frame
}
\examples{
sum_stat(mtcars, dplyr::everything())
}
