% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_names.R
\name{s_names}
\alias{s_names}
\title{Simple names}
\usage{
s_names(df)
}
\arguments{
\item{data.frame}{}
}
\value{
A data.frame with simple column names
}
\description{
s_names() removes symbols such as parentheses and colons and
substitute symbols (|, :, and +) with underscores. No letters are capitalized.
}
\examples{
x <- data.frame(x = c(1, 2), y = c(4, 2), z = c(1,2))
colnames(x) <- c("COL (1)", "col:2", "col-3")
x
s_names(x)
}
