% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stat.R
\name{sum_stat}
\alias{sum_stat}
\title{Summary statistics}
\usage{
sum_stat(df, variables)
}
\arguments{
\item{variables}{}
}
\value{
Summary statistics of selected variables
}
\description{
sum_stat returns summary statistics of selected variables
from a data frame
}
\examples{
var <- c("mpg", "cyl")
sum_stat(mtcars, var)
}
